New macro post-unwind-handler-bind, changed some instances of handler-bind to use it
The macro is used the same way as handler-bind, except that the stack
will be unwound before the handlers are executed. This means you cannot
restart from inside the body, but it also guarantees that unwind-protect
forms inside the body will be executed even if the handler terminates the
process. Handler-bind forms in antiweb.lisp that simply call fatal were
changed to reduce the chances of orphaning BDB locks.